requirements.txt
Flask==0.11
Flask-SQLAlchemy==2.1
flask-admin==1.4.2
psycopg2